add yml file to create conda environment
create a yml to create a conda environment (nlp) that has the packages
nessecary to run the literature clustering notebook in jupyterlab.

This file also includes the scispacy en_core_sci-lg model, which does
not need to be installed with a separate pip command.

note I spent some time troubleshooting getting this to work and ended up testingwith pip=10.0.1.